+++ pecst/ceramic/selection.py
@@ -7,7 +7,7 @@ import pathlib
 # 3rd party libraries
 import numpy as np
 import pandas as pd
-from matplotlib import pyplot as plt
+from matplotlib import pyplot as plt  # type: ignore
 
 # own libraries
 from pecst.cst_dataclasses import CapacitorRequirements
@@ -112,7 +112,7 @@ def select_ceramic_capacitors(c_requirements: CapacitorRequirements) -> tuple[li
 
     # sort out all capacitors where to many series capacitors are required (more than maximum in series allowed)
     ceramic_df["number_min_capacitors_in_series"] = np.ceil(
-        c_requirements.v_dc_for_op_max_voltage / (ceramic_df["voltage"] * (1 + c_requirements.voltage_safety_margin_percentage / 100)))
+        c_requirements.v_dc_for_op_max_voltage / (ceramic_df["voltage"] * (1 - c_requirements.voltage_safety_margin_percentage / 100)))
     logger.info(f"Initial ceramic dataframe shape: {ceramic_df.shape}")
     ceramic_df = ceramic_df.drop(ceramic_df[ceramic_df["number_min_capacitors_in_series"] > c_requirements.maximum_number_series_capacitors].index)
     logger.info(f"Ceramic dataframe shape after dropping number_min_capacitors_in_series > maximum_number_series_capacitors: {ceramic_df.shape}")
@@ -129,7 +129,7 @@ def select_ceramic_capacitors(c_requirements: CapacitorRequirements) -> tuple[li
                 x["ordering code"], x["capacitance"], v_dc_max, x["number_min_capacitors_in_series"], n_max_c, c_min_req), axis=1)
 
         # drop capacitors with no bias curve data given
-        # leads to drop all values ... but works with the foil capacitors
+        # leads to drop all values ... but works with the film capacitors
         # ceramic_df = ceramic_df.drop(ceramic_df[np.isnan(ceramic_df["in_series_needed"])].index)
         logger.info(f"Ceramic dataframe shape after dropping NaNs with in_series_needed: {ceramic_df.shape}")
         # loss calculation per capacitor