+++ pecst/foil/__init__.py
@@ -1,8 +1,0 @@
-"""Initialize foil capacitors."""
-from pecst.foil.current_capability import *
-from pecst.foil.dvdt import *
-from pecst.foil.esr_downloads import *
-from pecst.foil.lifetime import *
-from pecst.foil.power_loss import *
-from pecst.foil.read_capacitor_database import *
-from pecst.foil.selection import *