+++ pecst/functions.py
@@ -5,7 +5,7 @@ import logging
 
 # 3rd party libraries
 import numpy as np
-from matplotlib import pyplot as plt
+from matplotlib import pyplot as plt  # type: ignore
 
 # own libraries
 from pecst.cst_dataclasses import CapacitorRequirements, CalculatedRequirementsValues