+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2025-11-05
 ### Added
- - Initial release featuring foil capacitor selection for power electronics applications 
+ - Initial release featuring film capacitor selection for power electronics applications 
 
 [unreleased]: https://github.com/upb-lea/capacitor_selection_toolbox/compare/0.1.1...HEAD
 [0.1.1]: https://github.com/upb-lea/capacitor_selection_toolbox/compare/0.1.0...0.1.1